requirements.txt
# Lovable Remix Automation Dependencies
playwright==1.50.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.5
rich==13.9.4
